requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
intrinio-sdk==5.6.1
numpy==1.19.2
pandas==1.1.2
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
xtermcolor==1.3
